Moved the items use state and persist the data in the local storage to the app component.
