+++ src/App.jsx
@@ -1,17 +1,32 @@
-
+import React, { useState, useEffect } from 'react';
 import './App.css';
 import CreateItem from './components/CreateItem';
 import Header from './components/Header';
 import WelcomeScreen from './components/WelcomeScreen';
 
+const getDataFromLS = () => {
+    const data = localStorage.getItem('items');
+    if(data){
+        return JSON.parse(data)
+    }
+    else{
+        return []
+    }
+}
+
 function App() {
-  console.log(localStorage.length)
+  
+  const [items, setItems] = useState(getDataFromLS())
+
+    useEffect(() => {
+        localStorage.setItem('items',JSON.stringify(items))
+    }, [items])
+
   return (
     <div className="App">
       <Header />
 
-      {localStorage.length > 0 ? <CreateItem /> : <WelcomeScreen />}
-      
+      {items.length > 0 ? <CreateItem setItems={setItems} items={items} /> : <WelcomeScreen />}
       
     </div>
   );
